requirements.txt
segment-anything-py
opencv-python 
matplotlib 
pillow 
tqdm 
pandas
ultralytics  # YOLOv8 for training custom object detection models
torch
torchvision
pyyaml
requests
